requirements_mrcnn.txt
absl-py==0.15.0
alabaster==0.7.13
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==1.0.3
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
blobfile==3.0.0
cachetools==5.5.0
certifi==2019.11.28
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
comm==0.2.2
cryptography==2.8
curio==1.6
cycler==0.12.1
Cython==3.0.10
dbus-python==1.2.16
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distro==1.4.0
docrepr==0.2.0
docutils==0.20.1
entrypoints==0.3
exceptiongroup==1.2.1
executing==2.0.1
fasteners==0.14.1
fastjsonschema==2.19.1
filelock==3.16.0
flatbuffers==1.12
fonttools==4.51.0
fqdn==1.5.1
fsspec==2024.9.0
future==0.18.2
gast==0.4.0
google-auth==2.34.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.14.0
httpx==0.27.0
idna==2.8
imagesize==1.4.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipykernel==6.29.4
ipyparallel==8.8.0
ipython==7.34.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
Keras==2.1.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.5
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy_loader==0.4
linecache2==1.0.0
lockfile==0.12.2
lxml==5.3.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.6.3
matplotlib-inline==0.1.7
mistune==3.0.2
monotonic==1.5
mpmath==1.3.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.10.4
networkx==3.1
nose==1.3.7
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.19.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
opencv-python==4.9.0.80
opt-einsum==3.3.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.1
paramiko==2.6.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.6.0
pickleshare==0.7.5
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.1
pluggy==1.5.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.20.0
pycups==1.9.73
Pygments==2.17.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.1.2
pyRFC3339==1.1
pytest==7.0.1
pytest-asyncio==0.23.7
pytz==2024.1
PyWavelets==1.4.1
pyxdg==0.26
pyzmq==26.0.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.35.0
reportlab==3.5.34
requests==2.31.0
requests-oauthlib==2.0.0
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
scikit-image==0.16.2
scikit-learn==1.3.2
scipy==1.9.3
SecretStorage==2.3.1
Send2Trash==1.8.3
shapely==2.0.4
simplejson==3.16.0
six==1.15.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.1.2
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
ssh-import-id==5.10
stack-data==0.6.3
sympy==1.13.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
terminado==0.18.1
testpath==0.6.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tinycss2==1.3.0
tomli==2.0.1
torch==2.4.1
torchsummary==1.5.1
tornado==6.4
tqdm==4.66.4
traceback2==1.4.0
traitlets==5.14.3
trio==0.25.1
triton==3.0.0
types-python-dateutil==2.9.0.20240906
typing-extensions==3.7.4.3
tzdata==2024.1
unittest2==1.1.0
uri-template==1.3.0
urllib3==1.25.8
visdom==0.2.4
wadllib==1.3.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.10
wrapt==1.12.1
zipp==3.18.1
